refactor(pragma-cli): route kernel cross-domain imports through domain barrels
Sibling-domain reaches from src/kernel/** named individual files
(../<domain>/<file>.js) instead of the domain barrel #1026 added, which
cs:packaging.import.cross_domain and cs:packaging.export.barrel forbid.
Rewrites each to ../<domain>/index.js where the barrel already exports
every symbol the site imports; no barrel's export surface is widened.

Reroutes that would have put a lazily-imported module back on the
--help/__complete static graph are left alone: each candidate was applied
against the repo's own laziness guards (capabilities/lazy.test.ts,
kernel/completion/safety.test.ts) and reverted when it made zod, the
config reader, commander, @canonical/ke or runtime boot statically
reachable.
