+++ packages/cli/pragma/src/kernel/config/evaluateProjectConfig.ts
@@ -42,7 +42,7 @@ import {
 import { join } from "node:path";
 import { pathToFileURL } from "node:url";
 import { PROJECT_CONFIG_FILENAME, VERSION } from "../../constants.js";
-import { PragmaError } from "../error/PragmaError.js";
+import { PragmaError } from "../error/index.js";
 import { configCacheDir } from "./paths.js";
 import { parseRawConfig } from "./schema.js";
 import type { RawConfig } from "./types.js";
