+++ packages/cli/pragma/src/kernel/runtime/graphpack/read.ts
@@ -16,8 +16,7 @@ import { readFileSync } from "node:fs";
 import { join } from "node:path";
 import { createStore } from "@canonical/ke";
 import { compileFromExtraction } from "@canonical/ke-graphql";
-import { PragmaError } from "../../error/PragmaError.js";
-import { cliRecovery } from "../../error/recovery.js";
+import { cliRecovery, PragmaError } from "../../error/index.js";
 import type { StoreSession } from "../types.js";
 import { readManifest } from "./manifest.js";
 import { packIndexSchema } from "./schemas.js";
